% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkavro.R
\name{spark_write_avro}
\alias{spark_write_avro}
\title{Write a Spark DataFrame to a Avro file}
\usage{
spark_write_avro(x, path, mode = NULL, options = list())
}
\arguments{
\item{path}{The path to the file. Needs to be accessible from the cluster.
Supports the \samp{"hdfs://"}, \samp{"s3n://"} and \samp{"file://"} protocols.}

\item{mode}{Specifies the behavior when data or table already exists.}

\item{options}{A list of strings with additional options. See \url{http://spark.apache.org/docs/latest/sql-programming-guide.html#configuration}.}

\item{sc}{An active \code{spark_connection}.}

\item{name}{The name to assign to the newly generated table.}
}
\description{
Serialize a Spark DataFrame to the
\href{https://parquet.apache.org/}{Parquet} format.
}
